fix: Separate services between implementations and interfaces
