+++ src/main/java/school/faang/user_service/service/SkillOfferService.java
@@ -1,31 +1,16 @@
 package school.faang.user_service.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import school.faang.user_service.entity.recommendation.SkillOffer;
-import school.faang.user_service.repository.recommendation.SkillOfferRepository;
 
 import java.util.List;
 import java.util.Optional;
 
-@Service
-@RequiredArgsConstructor
-public class SkillOfferService {
-    private final SkillOfferRepository skillOfferRepository;
+public interface SkillOfferService {
+    List<SkillOffer> getSkillRequestsByRecommendationId(long recommendationId);
 
-    public List<SkillOffer> getSkillRequestsByRecommendationId(long recommendationId) {
-        return skillOfferRepository.findAllByRecommendationId(recommendationId);
-    }
+    void deleteSkillOfferssByRecommendationId(long recommendationId);
 
-    public void deleteSkillOfferssByRecommendationId(long recommendationId) {
-        skillOfferRepository.deleteAllByRecommendationId(recommendationId);
-    }
+    Optional<SkillOffer> findSkillOfferBySkillAndRecommendationIds(long skillId, long recommendationId);
 
-    public Optional<SkillOffer> findSkillOfferBySkillAndRecommendationIds(long skillId, long recommendationId) {
-        return skillOfferRepository.findBySkillIdAndRecommendationId(skillId, recommendationId);
-    }
-
-    public void createSkillOffer(long skillId, long recommendationId) {
-        skillOfferRepository.create(skillId, recommendationId);
-    }
+    void createSkillOffer(long skillId, long recommendationId);
 }
