+++ src/main/java/school/faang/user_service/service/UserServiceImpl.java
@@ -1,0 +1,53 @@
+package school.faang.user_service.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataRetrievalFailureException;
+import org.springframework.stereotype.Service;
+import school.faang.user_service.dto.CreateUserDto;
+import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.entity.User;
+import school.faang.user_service.exception.DataValidationException;
+import school.faang.user_service.exception.ExternalResourceNotFoundException;
+import school.faang.user_service.repository.UserRepository;
+import school.faang.user_service.service.avatarGenerator.AvatarGeneratorService;
+import school.faang.user_service.validator.CreateUserValidator;
+
+import java.io.IOException;
+
+@Service
+@RequiredArgsConstructor
+public class UserServiceImpl implements UserService {
+    private final UserRepository userRepository;
+
+    private final AvatarGeneratorService avatarGeneratorService;
+
+    private final CreateUserValidator createUserValidator;
+
+    @Override
+    public boolean doesUserExist(long userId) {
+        return userRepository.existsById(userId);
+    }
+
+    @Override
+    public User getUserById(long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new DataRetrievalFailureException(
+                        "User with id %d is not found".formatted(userId)));
+    }
+
+    @Override
+    public User findById(long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new DataValidationException("User not found"));
+    }
+
+    @Override
+    public boolean existsById(long userId) {
+        return userRepository.existsById(userId);
+    }
+
+    @Override
+    public UserDto createUser(CreateUserDto userDto) {
+        return null;
+    }
+}
