+++ src/test/java/school/faang/user_service/service/event/EventParticipationServiceImplTest.java
@@ -22,7 +22,6 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-
 @ExtendWith(MockitoExtension.class)
 class EventParticipationServiceImplTest {
     private static final long EVENT_ID = 1L;
